django_part/requirements.txt
attrs==16.1.0
autobahn==0.16.0
cffi==1.7.0
cryptography==1.5
cssselect==0.9.2
decorator==4.0.10
Django==1.10.1
django-redirect==0.2
idna==2.1
ipython==5.1.0
ipython-genutils==0.1.0
lxml==3.6.4
parsel==1.0.3
pexpect==4.2.1
pickleshare==0.7.4
pika==0.10.0
prompt-toolkit==1.0.7
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
Pygments==2.1.3
pyOpenSSL==16.1.0
pytest==3.0.2
python-scrapyd-api==2.0.1
queuelib==1.4.2
redis==2.10.5
requests==2.11.1
Scrapy==1.1.2
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
Twisted==16.4.0
txaio==2.5.1
w3lib==1.15.0
wcwidth==0.1.7
zope.interface==4.3.1
